server/requirements.txt
Django>=3.2,<3.3
djangorestframework==3.13.1
django-cors-headers==3.5.0
